move cluster migration to the operator, where it belongs
